requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.5
comm==0.2.2
curl_cffi==0.9.0
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.3.1
executing==2.2.0
frozendict==2.4.7
greenlet==3.1.1
idna==3.11
importlib_metadata==8.5.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.3.0
matplotlib-inline==0.1.7
multitasking==0.0.12
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
peewee==4.0.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.29.6
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.23
Pygments==2.19.1
PyMySQL==1.1.2
pyspark==3.5.8
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.2.1
requests==2.32.4
schedule==1.2.2
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.46
stack-data==0.6.3
tomli==2.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
yfinance==1.2.0
zipp==3.20.2
streamlit
plotly
streamlit-autorefresh
